chore(kit): `SliderKeyStepsDirective` all utils are in the directive's file
